refact: move room events to separated file
