+++ src/stores/room-store/index.ts
@@ -1,13 +1,14 @@
-import pusherJs, { Members, PresenceChannel } from "pusher-js";
+import pusherJs, { PresenceChannel } from "pusher-js";
 import createStore, { StateCreator } from "zustand";
-import { produce } from "immer";
+
+import { mountRoomEvents } from "./mount-room-events";
 import { api } from "../../lib/axios";
 import { createWebConnection } from "../../lib/pusher";
+
 import {
   BasicRoomInfo,
   EventMode,
   MainRoomEvents,
-  People,
   RoomInfo,
   RoomStoreProps,
 } from "./types";
@@ -18,134 +19,7 @@ const roomStore: StateCreator<RoomStoreProps, [], [], RoomStoreProps> = (
   set,
   get
 ) => {
-  const roomEvents = {
-    async onLoadPeople(people: { id: string; info: { name: string } }) {
-      const state = get();
-
-      set((state) => ({
-        peoples: state.peoples.concat({
-          id: people.id,
-          name: people.info.name,
-        }),
-      }));
-
-      const me = state.peoples.find(
-        (people) => people.id === state.basicInfo.subscription.members.myID
-      );
-
-      if (me.points !== undefined) {
-        await api.post("sync-people-points", {
-          senderPeople: {
-            id: me.id,
-            points: me.points,
-          },
-          targetPeopleID: people.id,
-        });
-      }
-    },
-
-    onPrepareRoom(roomPeoples: Members) {
-      const parsedMembers = Object.entries<People>(
-        roomPeoples.members
-      ).map<People>(([id, people]) => ({
-        id,
-        name: people.name,
-      }));
-
-      set({
-        peoples: parsedMembers,
-      });
-    },
-
-    onPeopleLeave(people: People) {
-      set((state) => ({
-        peoples: state.peoples.filter(
-          (statePeople) => statePeople.id !== people.id
-        ),
-      }));
-    },
-
-    onSelectPoint(people: Pick<People, "id" | "points">) {
-      set((state) => ({
-        peoples: state.peoples.map((statePeople) =>
-          statePeople.id === people.id
-            ? {
-                ...statePeople,
-                points: people.points,
-              }
-            : statePeople
-        ),
-      }));
-    },
-
-    onShowPoints({ show }: { show: boolean }) {
-      if (show) {
-        set(
-          produce((state: RoomStoreProps) => {
-            state.basicInfo.showPointsCountdown = 3;
-          })
-        );
-
-        const intervalID = setInterval(() => {
-          set(
-            produce((state: RoomStoreProps) => {
-              if (state.basicInfo.showPointsCountdown === 0) {
-                clearInterval(intervalID);
-                return;
-              }
-
-              state.basicInfo.showPointsCountdown--;
-            })
-          );
-        }, 1000);
-      }
-
-      set(
-        produce((state: RoomStoreProps) => {
-          state.basicInfo.showPoints = show;
-
-          if (!show) {
-            state.peoples = state.peoples.map((people) => ({
-              ...people,
-              points: undefined,
-            }));
-          }
-        })
-      );
-    },
-
-    onSyncPeoplePoints(senderPeople: { id: string; points: number }) {
-      set((state) => ({
-        peoples: state.peoples.map((people) =>
-          people.id === senderPeople.id
-            ? {
-                ...people,
-                points: senderPeople.points,
-              }
-            : people
-        ),
-      }));
-    },
-
-    onHighlightPeople({
-      sender_id: targetPeopleID,
-      highlight = true,
-    }: {
-      sender_id: string;
-      highlight?: boolean;
-    }) {
-      set((state) => ({
-        peoples: state.peoples.map((people) =>
-          people.id === targetPeopleID
-            ? {
-                ...people,
-                highlight,
-              }
-            : people
-        ),
-      }));
-    },
-  };
+  const roomEvents = mountRoomEvents(set, get);
 
   const INITIAL_STORE_VALUE: RoomStoreProps = {
     basicInfo: {
